moved more commands
