+++ src/action/buildCommands.ts
@@ -22,12 +22,6 @@ export default [
   /*
     Info-Text-Commands
     */
-  new SlashCommandBuilder().setName('metafrage')
-    .setDescription('Ein Text über Metafragen.'),
-  new SlashCommandBuilder().setName('codeblocks')
-    .setDescription('Ein Text über Codeblocks'),
-  new SlashCommandBuilder().setName('about')
-    .setDescription('About me'),
   /*
     Toggle-Role-Commands
     */
