+++ src/listeners/interactionCreate.ts
@@ -1,27 +1,18 @@
 import {
   CommandInteraction, Client, Interaction,
   ButtonInteraction, ApplicationCommandType, UserContextMenuCommandInteraction,
-  ModalSubmitInteraction, SelectMenuInteraction, MessageContextMenuCommandInteraction,
+  MessageContextMenuCommandInteraction,
   InteractionType, ChannelType, GuildBasedChannel,
   ActionRowBuilder, ButtonBuilder, ButtonStyle,
   AutocompleteInteraction
 } from 'discord.js'
 import { LogManager } from '../logger/logger'
-import { codeblocks, metafrage, about } from '../action/infoMessages'
 import { createRoleInterface } from '../action/roles_buttons_create'
-import { toggleRoles } from '../action/toggleRole'
 import startUserReport from '../action/userReport'
-
-import continueReport from '../action/continueReport'
-import finishReport from '../action/finishReport'
 import messageReport from '../action/messageReport'
 import { createTicket, ticketAdd, ticketClose } from '../action/ticket-system'
-import timeout from '../action/timeout'
-import { search } from '../action/google'
 import { AsyncDatabase } from '../sqlite/sqlite'
-import { createGiveaway, evalGiveaway, newParticipant } from '../action/giveaway'
 import { handleBlackJackCommands } from '../action/blackjack/handleCommands'
-import rename from '../action/rename'
 // import { autocomplete } from "../action/youtube";
 
 export default (client: Client, logger: LogManager, db: AsyncDatabase): void => {
@@ -38,12 +29,6 @@ export default (client: Client, logger: LogManager, db: AsyncDatabase): void => 
     if (interaction.isContextMenuCommand() && interaction.commandType === ApplicationCommandType.Message) {
       await handleMessageContextMenuCommand(client, interaction, logger, db)
     }
-    if (interaction.isStringSelectMenu()) {
-      await handleSelectMenu(client, interaction, logger, db)
-    }
-    if (interaction.isModalSubmit()) {
-      await handleModalSubmit(client, interaction, logger, db)
-    }
     if (interaction.type === InteractionType.ApplicationCommandAutocomplete) {
       await handleAutoComplete(client, interaction, logger, db)
     }
@@ -54,29 +39,11 @@ const handleSlashCommand = async (client: Client, interaction: CommandInteractio
   let channel: GuildBasedChannel | null
   // handle slash command here
   switch (interaction.commandName) {
-    case 'metafrage':
-      await metafrage(client, interaction, logger.logger('Metafrage'))
-      return
-    case 'codeblocks':
-      await codeblocks(client, interaction, logger.logger('Codeblocks'))
-      return
     case 'role':
       await createRoleInterface(interaction, 'once', logger.logger('Toogle-Roles'))
       return
     case 'role-force-button':
       await createRoleInterface(interaction, 'global', logger.logger('Toggle-Roles'))
-      return
-    case 'about':
-      await about(client, interaction, logger.logger('About'))
-      return
-    case 'rename':
-      await rename(client, interaction, logger.logger('Rename'))
-      return
-    case 'search':
-      await search(client, interaction, logger.logger('search'))
-      return
-    case 'timeout':
-      await timeout(client, interaction, logger.logger('timeout'), db)
       return
     case 'ticket-create':
       if ((interaction.guild == null) || (interaction.member == null) || !interaction.member.user.id) {
@@ -134,12 +101,6 @@ const handleSlashCommand = async (client: Client, interaction: CommandInteractio
         ephemeral: true
       })
       break
-    case 'giveaway':
-      await createGiveaway(client, interaction, db)
-      break
-    case 'giveaway-eval':
-      await evalGiveaway(client, interaction, db)
-      break
     case 'bj':
       await handleBlackJackCommands(interaction, logger)
       break
@@ -147,12 +108,6 @@ const handleSlashCommand = async (client: Client, interaction: CommandInteractio
 }
 
 const handleButtonInteraction = async (client: Client, interaction: ButtonInteraction, db: AsyncDatabase, logger: LogManager): Promise<void> => {
-  if (/(addRole-|removeRole-)([0-9]+)/.test(interaction.customId)) {
-    await toggleRoles(client, interaction, logger.logger('Toggle-Roles'))
-  }
-  if (interaction.customId === 'giveaway-participate') {
-    await newParticipant(client, interaction, db)
-  }
   if (interaction.customId === 'ticket-delete') {
     await interaction.reply({
       content: 'Lösche Ticket',
@@ -185,18 +140,6 @@ const handleUserContextMenuCommand = async (client: Client, interaction: UserCon
 const handleMessageContextMenuCommand = async (client: Client, interaction: MessageContextMenuCommandInteraction, logger: LogManager, db: AsyncDatabase): Promise<void> => {
   if (interaction.commandType === ApplicationCommandType.Message && interaction.commandName === 'REPORT') {
     await messageReport(interaction, client, db, logger.logger('Report-System'))
-  }
-}
-
-const handleSelectMenu = async (client: Client, interaction: SelectMenuInteraction, logger: LogManager, db: AsyncDatabase): Promise<void> => {
-  if (/report_.+_category/.test(interaction.customId)) {
-    await continueReport(interaction, client, db, logger.logger('Report'))
-  }
-}
-
-const handleModalSubmit = async (client: Client, interaction: ModalSubmitInteraction, logger: LogManager, db: AsyncDatabase): Promise<void> => {
-  if (/report_.+_finish/.test(interaction.customId)) {
-    await finishReport(interaction, client, db, logger.logger('Report'))
   }
 }
