+++ src/modalInteractions/report_finish.ts
@@ -1,0 +1,113 @@
+import { ChannelType, Colors, EmbedBuilder, ModalSubmitInteraction } from "discord.js"
+import LogManager from "src/logger/logger"
+import { AsyncDatabase } from "src/sqlite/sqlite"
+
+export default {
+  customId: /report_.+_finish/,
+  async execute (interaction: ModalSubmitInteraction): Promise<void> {
+    const logger = LogManager.getInstance().logger('Report-System')
+    const db = await AsyncDatabase.open()
+    if (!db) {
+      logger.logSync('ERROR', 'DB konnte nicht geöffnet werden.')
+      return
+    }
+    const uuid = interaction.customId.split('_')[1]
+    console.log(`Report ${interaction.customId} ${uuid}`)
+
+    let result
+    try {
+      result = await db.getAsync('SELECT uuid, reported_id, status, category, message FROM reports WHERE uuid = ?', [uuid])
+    } catch (err) {
+      await interaction.reply('ID existiert nicht. Bitte mache den Report von neu.')
+      return
+    }
+
+    await db.runAsync('UPDATE reports SET description = ? WHERE uuid = ?', [
+      interaction.fields.getTextInputValue('description'), uuid
+    ])
+
+    await interaction.reply({
+      content: 'Dein Report wurde an das Mod-Team übermittelt. Du kannst im neu erstellten Ticket Datails, Screenshots und ähnliches teilen.',
+      embeds: [
+        new EmbedBuilder()
+          .setAuthor({
+            name: '🚩 REPORT-SYSTEM'
+          })
+          .setColor(Colors.Red)
+          .setTitle('Report')
+          .setDescription(interaction.fields.getTextInputValue('description'))
+          .addFields(
+            {
+              name: 'Absender',
+              value: `${interaction.member?.user.username as string}#${interaction.member?.user.discriminator as string} <@${interaction.member?.user.id as string}> #${interaction.member?.user.id as string}`
+            },
+            {
+              name: 'Beschuldigt',
+              value: (await interaction.guild?.members.fetch(result?.reported_id ?? '') ?? '<ERROR>').toString()
+            },
+            {
+              name: 'Regel',
+              value: result?.category ?? '<ERROR>'
+            },
+            {
+              name: 'ID',
+              value: result?.uuid ?? '<ERROR>'
+            },
+            {
+              name: 'Status',
+              value: 'Fertig'
+            },
+            {
+              name: 'Verlinkte Nachricht',
+              value: result.message ?? '<_NULL_>'
+            }
+          )
+      ],
+      ephemeral: true
+    })
+
+    const channel = await interaction.guild?.channels.fetch(process.env.REPORT_CHANNEL_ID ?? '')
+    if ((channel == null) || channel?.type !== ChannelType.GuildText) {
+      logger.logSync('ERROR', 'Channel not found / not TextBased')
+      return
+    }
+    await channel.send({
+      content: 'Ein Report wurde erstellt.',
+      embeds: [
+        new EmbedBuilder()
+          .setAuthor({
+            name: '🚩 REPORT-SYSTEM'
+          })
+          .setColor(Colors.Red)
+          .setTitle('Report')
+          .setDescription(interaction.fields.getTextInputValue('description'))
+          .addFields(
+            {
+              name: 'Absender',
+              value: interaction.member?.toString() ?? ''
+            },
+            {
+              name: 'Beschuldigt',
+              value: (await interaction.guild?.members.fetch(result?.reported_id ?? '') ?? '<ERROR>').toString()
+            },
+            {
+              name: 'Regel',
+              value: result?.category ?? '<ERROR>'
+            },
+            {
+              name: 'ID',
+              value: result?.uuid ?? '<ERROR>'
+            },
+            {
+              name: 'Status',
+              value: 'Fertig'
+            },
+            {
+              name: 'Verlinkte Nachricht',
+              value: result?.message ?? '<NULL>'
+            }
+          )
+      ]
+    })
+  }
+}
